+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - PowerPoint2007 Writer : Bugfix for opening PPTX on Mac - @thsteinmetz GH-89
 
 ### Changes
+- PhpOffice\PhpPresentation\getProperties has been deprecated for PhpOffice\PhpPresentation\getDocumentProperties - @Progi1984 GH-154
+- PhpOffice\PhpPresentation\setProperties has been deprecated for PhpOffice\PhpPresentation\setDocumentProperties - @Progi1984 GH-154
 - PhpOffice\PhpPowerpoint\Style\Alignment::setLevel can now be defined great than 8 - @Progi1984 GH-141
 
 ### Features
@@ -16,6 +18,7 @@
 - PowerPoint2007 Reader/Writer : Mark as final - @Progi1984 GH-118
 - PowerPoint2007 Reader/Writer : Set default zoom value for presentation - @Progi1984 GH-122
 - PowerPoint2007 Reader/Writer : Slide Background Color or Image - @Progi1984 GH-152
+- PowerPoint2007 Reader/Writer : Add Properties for allowing loop continuously until 'Esc' - @Progi1984 GH-154
 
 ## 0.5.0 - 2015-10-08